Extension.ts modularized
