+++ webview-ui/src/app/app.component.html
@@ -1,10 +1,3 @@
-<!-- <div class="buttonContainer">
-  <div>
-    <button class="button-77" (click)="toggleView()">
-      {{ currentView === 'folder-file' ? 'Parent Child View' : 'File System View' }}
-    </button>
-  </div>
-</div> -->
 
 <div class="buttonContainer">
   <div>
