+++ webview-ui/src/app/app.component.ts
@@ -12,56 +12,32 @@ export class AppComponent implements OnInit {
   ngOnInit() {
     //start view as the folder-file hierarchy graph
     this.currentView = 'folder-file';
-
-    //add listener to catch and store image URI's to pass down to view components to load icons
-    // window.addEventListener('message', (event) => {
-    //   const message: ExtensionMessage = event.data;
-
-    //   switch(message.command) {
-    //     case 'updateUris': {
-    //       this.uris = message.data;
-    //       break;
-    //     }
-
-    //     default:
-    //       console.log('Unknown in app event listener -> ', message.command);
-    //       break;
-    //   }
-    // });
   }
 
-  //toggle visible divs on button click, see app.html
-  toggleView() {
-    this.currentView =
-      this.currentView === 'folder-file' ? 'parent-child' : 'folder-file';
-    if (this.currentView === 'folder-file') {
+  loadFolderFile() {
+    if (this.currentView === 'parent-child') {
+
+      this.currentView = 'folder-file';
+      
       vscode.postMessage({
         command: 'reloadFolderFile',
         data: {},
       });
+    
     }
-    if (this.currentView === 'parent-child') {
+  }
+
+  loadParentChild() {
+    if (this.currentView === 'folder-file') {
+
+      this.currentView = 'parent-child';
+
       vscode.postMessage({
         command: 'loadParentChild',
         data: {},
       });
+    
     }
-  }
-
-  loadFolderFile() {
-    this.currentView = 'folder-file';
-    vscode.postMessage({
-      command: 'reloadFolderFile',
-      data: {},
-    });
-  }
-
-  loadParentChild() {
-    this.currentView = 'parent-child';
-    vscode.postMessage({
-      command: 'loadParentChild',
-      data: {},
-    });
   }
 
   // nodes, edges, uris
